% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgeDataCode.R
\name{build_ages}
\alias{build_ages}
\title{Get Age data from PBDB to pair to the tips of trees}
\usage{
build_ages(data_path, xml_path, out_path)
}
\arguments{
\item{data_path}{A path to nexus datafiles}

\item{xml_path}{A path to xml metadata for each dataset}

\item{outpath}{Where you'd like to write the data to.}
}
\value{
AgesMatrix A dataframe of taxa and their minimum and maximum occurrence
}
\description{
For each nexus dataset, locate an XML file containing ages. If any are missing, query PBDB to see if they can be found. Returns a TSV-file of ages per tip.
}
