% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomPlots.R
\name{custom_plots}
\alias{custom_plots}
\title{Create treespace ordination including extra columns.}
\usage{
custom_plots(CombinedTrees, StratFile, ext_columns = c("MIG", "Bayes"))
}
\arguments{
\item{CombinedTrees}{A list of all the parsimony and Bayesian trees you'd like to plot}

\item{StratFile}{The logfile containing your stratigraphic congrunence values. Should contain a column to indicate if the tree is Bayesian or not.}

\item{ext_columns}{Any extra columns you'd like to color your plot by}

\item{n.points}{The number of points on each plot}

\item{fill.color}{The name of any column in your parameter file that you would like to use as a fill colour for the points of the plot.}
}
\value{
A matrix of ordinations in treespace.
}
\description{
This function will take list of rwty.chains objects and produce a matrix of ordinations in treespace based on Robinson-Foulds Distance. The points can be colored and sized by extra columns listed in the `ext_columns` argument. This function is modified from the `RWTY` package.
}
