% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_treespace.R
\name{custom_treespace}
\alias{custom_treespace}
\title{Create treespace ordination including extra columns.}
\usage{
custom_treespace(chains, n.points = 100, burnin = 0, fill.color = NA,
  ext_columns = NULL)
}
\arguments{
\item{chains}{A list of one or more rwty.chain objects}

\item{n.points}{The number of points on each plot}

\item{burnin}{The number of samples to remove from the start of the chain as burnin}

\item{fill.color}{The name of any column in your parameter file that you would like to use as a fill colour for the points of the plot.}

\item{ext_columns}{Any extra columns you'd like to color your plot by}
}
\value{
A matrix of ordinations in treespace.
}
\description{
This function will take list of rwty.chains objects and produce a matrix of ordinations in treespace based on Robinson-Foulds Distance. The points can be colored and sized by extra columns listed in the `ext_columns` argument.
}
