% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StratCongruenceCode.R
\name{make_strat_congruence}
\alias{make_strat_congruence}
\title{R script to perform strat congruence tests.
This R function runs statigraphic congruence tests on both a Bayesian posterior sample and a set of most parsimonious trees. Nothing is returned, but boxplots of these figures are saved to a results directory.}
\usage{
make_strat_congruence(RevBayesFiles, NexusFiles, AgeFiles, ParsimonyFiles,
  RB_output, Parsimony_output)
}
\description{
R script to perform strat congruence tests.
This R function runs statigraphic congruence tests on both a Bayesian posterior sample and a set of most parsimonious trees. Nothing is returned, but boxplots of these figures are saved to a results directory.
}
