% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplots.R
\name{makeBoxplot}
\alias{makeBoxplot}
\title{Create box plots of our data}
\usage{
makeBoxplot(x)
}
\arguments{
\item{x}{file containing stratigraphic congruence metrics}
}
\description{
This function will take a file of stratigraphic congrunce metrics and make a box plot. Returns nothing.
}
