% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/violin_plots.R
\name{makeViolins}
\alias{makeViolins}
\title{Create violin plots of our data}
\usage{
makeViolins(x)
}
\arguments{
\item{x}{file containing stratigraphic congruence metrics}
}
\description{
This function will take a file of stratigraphic congrunce metrics and make a violin plot. Returns nothing.
}
