% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteringCode.R
\name{remove_missing_tips}
\alias{remove_missing_tips}
\title{Custom functions to filter datasets by completeness of tip sampling}
\usage{
remove_missing_tips(xml_path)
}
\arguments{
\item{datasets}{A vector of datasets that you would like to clean of too-small datasets or datasets with tip ages missing. These must be xml metadata files.}
}
\value{
vector A vector datasets with the datasets with too few tips or missing tips removed
}
\description{
Some datasets in PBDB have unreconciled lack full tip sampling, or have fewer than five tips. These functions allow us to remove those datasets.
}
