% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilteringCode.R
\name{remove_unreconciled}
\alias{remove_unreconciled}
\title{Custom functions to filter datasets down to those that do not have unreconciled taxa}
\usage{
remove_unreconciled(xml_path)
}
\arguments{
\item{xml_path}{A directory of datasets that you would like to clean of unreconiled taxa. These must be xml metadata files.}
}
\value{
vector A vector datasets with the unreconciled taxa removed
}
\description{
Some datasets in PBDB have unreconciled taxa names. We remove these before 
proceding.
}
