% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupAll.R
\name{summarize_medians}
\alias{summarize_medians}
\title{Process MIG score for all the files used in this study.}
\usage{
summarize_medians(user.path = NULL, extension = NULL)
}
\arguments{
\item{extension}{Extension of the files to process}

\item{path}{relative path to your data}
}
\value{
median_tbl A table of median values for MIG across datasets.
}
\description{
This function reads in the output of Strap's MIG claculations. It can read in one file, or many, provided a 
relative path to your files and the extension of the files. This function then prcoesses the files to retrieve 
the median MIG scores for both Bayesian and parsimony analyses.
}
