% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryStats.R
\name{summary_stat}
\alias{summary_stat}
\title{Generate the S1 table of Wright and Lloyd 2019}
\usage{
summary_stat(dataf)
}
\arguments{
\item{dataf}{The file you'd like to examine}

\item{measure}{Which stratigraphic congruence measure you'd like to look at}

\item{by_col}{If you'd like the results grouped. Defaults to a column called "Bayes"}
}
\value{
summary a data frame showing mean scores for your metric of choice, grouped by the by_col col
}
\description{
This function takes as its arguments what straitgraphic congruence measure you want
from a file of dataset of stratigraphic congruence metrics. Assumes by default that you want to group by the column that says whether the analysis was Bayesian or not.
}
